+++ adapters/config.go
@@ -1,4 +1,4 @@
-package repository
+package adapters
 
 import (
 	"os"
