+++ cmd/migrate/main.go
@@ -3,19 +3,19 @@ package main
 import (
 	"fmt"
 
+	"github.com/ditointernet/tradulab-service/adapters"
 	"github.com/ditointernet/tradulab-service/driven"
-	"github.com/ditointernet/tradulab-service/repository"
 )
 
 func main() {
 
-	env, pan := repository.GoDotEnvVariable()
+	env, pan := adapters.GoDotEnvVariable()
 	if pan != nil {
 		fmt.Println("Error during environment variables build", pan.Error())
 		return
 	}
 
-	db := repository.NewDatabase(&repository.Config{
+	db := adapters.NewDatabase(&adapters.Config{
 		User:     env.User,
 		Host:     env.Host,
 		Password: env.Password,
