+++ internal/core/ports/repository.go
@@ -1,7 +1,9 @@
 package ports
 
-import "github.com/ditointernet/tradulab-service/driven"
+import (
+	"github.com/ditointernet/tradulab-service/internal/core/domain"
+)
 
 type Repository interface {
-	SaveFile(*driven.File) error
+	SaveFile(*domain.File) error
 }
